add docker file for go development
