--- lambda/welcomeBoard/go.mod
@@ -1,5 +1,0 @@
-module welcomeBoard/main
-
-go 1.18
-
-require github.com/aws/aws-lambda-go v1.32.1 // indirect
