--- lambda/welcomeBoard/main.go
@@ -1,19 +1,0 @@
-package main
-
-import (
-        "fmt"
-        "context"
-        "github.com/aws/aws-lambda-go/lambda"
-)
-
-type MyEvent struct {
-        Name string `json:"name"`
-}
-
-func HandleRequest(ctx context.Context, name MyEvent) (string, error) {
-        return fmt.Sprintf("Hello %s!", name.Name ), nil
-}
-
-func main() {
-        lambda.Start(HandleRequest)
-}
