+++ lambda/welcomeBoard/src/go.mod
@@ -1,0 +1,3 @@
+module wedding.com/main
+
+go 1.19
